requirements.txt
# Adizon Dependencies - Updated for LangGraph (Dec 2025)

# === Core Framework ===
fastapi==0.115.5
uvicorn==0.32.1
starlette==0.41.3
python-dotenv==1.0.1

# === LangChain & LangGraph ===
langchain==0.3.13
langchain-core==0.3.63
langchain-community==0.3.13
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
langsmith==0.2.11
langgraph>=0.2.60
langgraph-checkpoint-postgres>=2.0.0

# === Database (PostgreSQL) ===
SQLAlchemy==2.0.45
alembic==1.13.1
psycopg[binary,pool]>=3.2.0

# === Pydantic V2 ===
pydantic==2.10.4
pydantic-settings==2.12.0
pydantic_core==2.27.2
annotated-types==0.7.0
email-validator==2.2.0

# === HTTP & Networking ===
httpx==0.28.1
httpcore==1.0.9
httpx-sse==0.4.3
requests==2.32.3
requests-toolbelt==1.0.0
aiohttp==3.13.2
aiohappyeyeballs==2.6.1
aiosignal==1.4.0

# === Chat Platforms ===
slack-sdk==3.27.0

# === AI/ML ===
openai==1.109.1
tiktoken==0.12.0
numpy==2.4.0

# === Utilities ===
PyYAML==6.0.3
rapidfuzz==3.10.1
orjson==3.11.5
click==8.3.1
tqdm==4.67.1
tenacity==9.1.2

# === Serialization ===
dataclasses-json==0.6.7
marshmallow==3.26.2
jsonpatch==1.33
jsonpointer==3.0.0

# === Type Checking ===
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
mypy_extensions==1.1.0

# === Other ===
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.2
anyio==4.12.0
sniffio==1.3.1
attrs==25.4.0
frozenlist==1.8.0
multidict==6.7.0
propcache==0.4.1
yarl==1.22.0
jiter==0.12.0
regex==2025.11.3
distro==1.9.0
h11==0.16.0
packaging==24.2
